requirements.txt
maxapi-python==1.1.20
python-telegram-bot==20.0